+++ app/setup/index.js
@@ -3,7 +3,7 @@ const Disconnect = require('disconnect');
 const winston = require('winston');
 const SpotifyWebApi = require('spotify-web-api-node');
 
-const { store, actions } = require('../src/redux/state');
+const { store, actions } = require('../src/state');
 const createSearchFunction = require('../src/search');
 const createSpotifyModule = require('../src/api/spotify');
 const Query = require('../src/query');
